refactor: P2-T0 - 预拆分播放相关 ViewModel 代码
